+++ twenty-first/src/storage/storage_vec/mod.rs
@@ -1,522 +1,26 @@
-use super::level_db::DB;
-use itertools::Itertools;
-use leveldb::{batch::WriteBatch, options::ReadOptions};
-use serde::{de::DeserializeOwned, Serialize};
-use std::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
-use std::{
-    collections::{HashMap, VecDeque},
-    sync::Arc,
-};
-
-pub type Index = u64;
-
-pub trait StorageVec<T> {
-    /// check if collection is empty
-    fn is_empty(&self) -> bool;
-
-    /// get collection length
-    fn len(&self) -> Index;
-
-    /// get single element at index
-    fn get(&self, index: Index) -> T;
-
-    /// get multiple elements matching indices
-    fn get_many(&self, indices: &[Index]) -> Vec<T>;
-
-    /// get all elements
-    fn get_all(&self) -> Vec<T>;
-
-    /// set a single element.
-    fn set(&mut self, index: Index, value: T);
-
-    /// set multiple elements.
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>);
-
-    /// set all elements with a simple list of values in an array or Vec.
-    ///
-    /// calls ::set_many() internally.
-    ///
-    /// panics if input length does not match target length.
-    ///
-    /// note: casts the array's indexes from usize to Index.
-    fn set_all(&mut self, vals: &[T])
-    where
-        T: Clone,
-    {
-        assert!(
-            vals.len() as Index == self.len(),
-            "size-mismatch.  input has {} elements and target has {} elements.",
-            vals.len(),
-            self.len(),
-        );
-
-        self.set_many(
-            vals.iter()
-                .enumerate()
-                .map(|(i, v)| (i as Index, v.clone())),
-        );
-    }
-
-    /// pop an element from end of collection
-    fn pop(&mut self) -> Option<T>;
-
-    /// push an element to end of collection
-    fn push(&mut self, value: T);
-}
-
-pub enum WriteElement<T: Serialize + DeserializeOwned> {
-    OverWrite((Index, T)),
-    Push(T),
-    Pop,
-}
-
-/// This is the private impl of RustyLevelDbVec.
-///
-/// RustyLevelDbVec is a public wrapper that adds RwLock around
-/// all accesses to RustyLevelDbVecPrivate
-pub(crate) struct RustyLevelDbVecPrivate<T: Serialize + DeserializeOwned> {
-    key_prefix: u8,
-    db: Arc<DB>,
-    write_queue: VecDeque<WriteElement<T>>,
-    length: Index,
-    cache: HashMap<Index, T>,
-    name: String,
-}
-
-impl<T: Serialize + DeserializeOwned + Clone> StorageVec<T> for RustyLevelDbVecPrivate<T> {
-    fn is_empty(&self) -> bool {
-        self.length == 0
-    }
-
-    fn len(&self) -> Index {
-        self.length
-    }
-
-    fn get(&self, index: Index) -> T {
-        // Disallow getting values out-of-bounds
-        assert!(
-            index < self.len(),
-            "Out-of-bounds. Got {index} but length was {}. persisted vector name: {}",
-            self.length,
-            self.name
-        );
-
-        // try cache first
-        if self.cache.contains_key(&index) {
-            return self.cache[&index].clone();
-        }
-
-        // then try persistent storage
-        let db_key = self.get_index_key(index);
-        let db_val = self
-            .db
-            .get_u8(&ReadOptions::new(), &db_key)
-            .unwrap_or_else(|_| {
-                panic!(
-                    "Element with index {index} does not exist in {}. This should not happen",
-                    self.name
-                )
-            })
-            .unwrap();
-        bincode::deserialize(&db_val).unwrap()
-    }
-
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        fn sort_to_match_requested_index_order<T>(indexed_elements: HashMap<usize, T>) -> Vec<T> {
-            let mut elements = indexed_elements.into_iter().collect_vec();
-            elements.sort_unstable_by_key(|&(index_position, _)| index_position);
-            elements.into_iter().map(|(_, element)| element).collect()
-        }
-
-        assert!(
-            indices.iter().all(|x| *x < self.len()),
-            "Out-of-bounds. Got indices {indices:?} but length was {}. persisted vector name: {}",
-            self.len(),
-            self.name
-        );
-
-        let (indices_of_elements_in_cache, indices_of_elements_not_in_cache): (Vec<_>, Vec<_>) =
-            indices
-                .iter()
-                .copied()
-                .enumerate()
-                .partition(|&(_, index)| self.cache.contains_key(&index));
-
-        let mut fetched_elements = HashMap::with_capacity(indices.len());
-        for (index_position, index) in indices_of_elements_in_cache {
-            let element = self.cache[&index].clone();
-            fetched_elements.insert(index_position, element);
-        }
-
-        let no_need_to_lock_database = indices_of_elements_not_in_cache.is_empty();
-        if no_need_to_lock_database {
-            return sort_to_match_requested_index_order(fetched_elements);
-        }
-
-        // let db_reader = self.db;
-
-        let elements_fetched_from_db = indices_of_elements_not_in_cache
-            .iter()
-            .map(|&(_, index)| self.get_index_key(index))
-            .map(|key| self.db.get_u8(&ReadOptions::new(), &key).unwrap().unwrap())
-            .map(|db_element| bincode::deserialize(&db_element).unwrap());
-
-        let indexed_fetched_elements_from_db = indices_of_elements_not_in_cache
-            .iter()
-            .map(|&(index_position, _)| index_position)
-            .zip_eq(elements_fetched_from_db);
-        fetched_elements.extend(indexed_fetched_elements_from_db);
-
-        sort_to_match_requested_index_order(fetched_elements)
-    }
-
-    /// Return all stored elements in a vector, whose index matches the StorageVec's.
-    /// It's the caller's responsibility that there is enough memory to store all elements.
-    fn get_all(&self) -> Vec<T> {
-        let length = self.len();
-
-        let (indices_of_elements_in_cache, indices_of_elements_not_in_cache): (Vec<_>, Vec<_>) =
-            (0..length).partition(|index| self.cache.contains_key(index));
-
-        let mut fetched_elements: Vec<Option<T>> = vec![None; length as usize];
-        for index in indices_of_elements_in_cache {
-            let element = self.cache[&index].clone();
-            fetched_elements[index as usize] = Some(element);
-        }
-
-        let no_need_to_lock_database = indices_of_elements_not_in_cache.is_empty();
-        if no_need_to_lock_database {
-            return fetched_elements
-                .into_iter()
-                .map(|x| x.unwrap())
-                .collect_vec();
-        }
-
-        // let db_reader = self.db;
-        for index in indices_of_elements_not_in_cache {
-            let key = self.get_index_key(index);
-            let element = self.db.get_u8(&ReadOptions::new(), &key).unwrap().unwrap();
-            let element = bincode::deserialize(&element).unwrap();
-            fetched_elements[index as usize] = Some(element);
-        }
-
-        fetched_elements
-            .into_iter()
-            .map(|x| x.unwrap())
-            .collect_vec()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        // Disallow setting values out-of-bounds
-        assert!(
-            index < self.len(),
-            "Out-of-bounds. Got {index} but length was {}. persisted vector name: {}",
-            self.length,
-            self.name
-        );
-
-        if let Some(_old_val) = self.cache.insert(index, value.clone()) {
-            // If cache entry exists, we remove any corresponding
-            // OverWrite ops in the `write_queue` to reduce disk IO.
-
-            // logic: retain all ops that are not overwrite, and
-            // overwrite ops that do not have an index matching cache_index.
-            self.write_queue.retain(|op| match op {
-                WriteElement::OverWrite((i, _)) => *i != index,
-                _ => true,
-            })
-        }
-
-        self.write_queue
-            .push_back(WriteElement::OverWrite((index, value)));
-    }
-
-    /// set multiple elements.
-    ///
-    /// panics if key_vals contains an index not in the collection
-    ///
-    /// It is the caller's responsibility to ensure that index values are
-    /// unique.  If not, the last value with the same index will win.
-    /// For unordered collections such as HashMap, the behavior is undefined.
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        for (index, value) in key_vals.into_iter() {
-            self.set(index, value);
-        }
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        // add to write queue
-        self.write_queue.push_back(WriteElement::Pop);
-
-        // If vector is empty, return None
-        if self.length == 0 {
-            return None;
-        }
-
-        // Update length
-        self.length -= 1;
-
-        // try cache first
-        if self.cache.contains_key(&self.length) {
-            self.cache.remove(&self.length)
-        } else {
-            // then try persistent storage
-            let db_key = self.get_index_key(self.length);
-            self.db
-                .get_u8(&ReadOptions::new(), &db_key)
-                .unwrap()
-                .map(|bytes| bincode::deserialize(&bytes).unwrap())
-        }
-    }
-
-    fn push(&mut self, value: T) {
-        // add to write queue
-        self.write_queue
-            .push_back(WriteElement::Push(value.clone()));
-
-        // record in cache
-        let _old_value = self.cache.insert(self.length, value);
-
-        // note: we cannot naively remove any previous `Push` ops with
-        // this value from the write_queue (to reduce disk i/o) because
-        // there might be corresponding `Pop` op(s).
-
-        // update length
-        self.length += 1;
-    }
-}
-
-impl<T: Serialize + DeserializeOwned> RustyLevelDbVecPrivate<T> {
-    // Return the key used to store the length of the persisted vector
-    fn get_length_key(key_prefix: u8) -> [u8; 2] {
-        const LENGTH_KEY: u8 = 0u8;
-        [key_prefix, LENGTH_KEY]
-    }
-
-    /// Return the length at the last write to disk
-    fn persisted_length(&self) -> Index {
-        let key = Self::get_length_key(self.key_prefix);
-        match self.db.get_u8(&ReadOptions::new(), &key).unwrap() {
-            Some(value) => bincode::deserialize(&value).unwrap(),
-            None => 0,
-        }
-    }
-
-    /// Return the level-DB key used to store the element at an index
-    fn get_index_key(&self, index: Index) -> [u8; 9] {
-        [vec![self.key_prefix], bincode::serialize(&index).unwrap()]
-            .concat()
-            .try_into()
-            .unwrap()
-    }
-
-    pub fn new(db: Arc<DB>, key_prefix: u8, name: &str) -> Self {
-        let length_key = Self::get_length_key(key_prefix);
-        let length = match db.get_u8(&ReadOptions::new(), &length_key).unwrap() {
-            Some(length_bytes) => bincode::deserialize(&length_bytes).unwrap(),
-            None => 0,
-        };
-        let cache = HashMap::new();
-        Self {
-            key_prefix,
-            db,
-            write_queue: VecDeque::default(),
-            length,
-            cache,
-            name: name.to_string(),
-        }
-    }
-
-    /// Collect all added elements that have not yet bit persisted
-    pub fn pull_queue(&mut self, write_batch: &mut WriteBatch) {
-        let original_length = self.persisted_length();
-        let mut length = original_length;
-        while let Some(write_element) = self.write_queue.pop_front() {
-            match write_element {
-                WriteElement::OverWrite((i, t)) => {
-                    let key = self.get_index_key(i);
-                    let value = bincode::serialize(&t).unwrap();
-                    write_batch.put_u8(&key, &value);
-                }
-                WriteElement::Push(t) => {
-                    let key =
-                        [vec![self.key_prefix], bincode::serialize(&length).unwrap()].concat();
-                    length += 1;
-                    let value = bincode::serialize(&t).unwrap();
-                    write_batch.put(&key, &value);
-                }
-                WriteElement::Pop => {
-                    let key = [
-                        vec![self.key_prefix],
-                        bincode::serialize(&(length - 1)).unwrap(),
-                    ]
-                    .concat();
-                    length -= 1;
-                    write_batch.delete(&key);
-                }
-            };
-        }
-
-        if original_length != length {
-            let key = Self::get_length_key(self.key_prefix);
-            write_batch.put_u8(&key, &bincode::serialize(&self.length).unwrap());
-        }
-
-        self.cache.clear();
-    }
-}
-
-/// RustyLevelDbVec is a concurrency safe database-backed
-/// Vec with in memory read/write caching for all operations.
-pub struct RustyLevelDbVec<T: Serialize + DeserializeOwned> {
-    inner: Arc<RwLock<RustyLevelDbVecPrivate<T>>>,
-}
-
-impl<T: Serialize + DeserializeOwned + Clone> StorageVec<T> for RustyLevelDbVec<T> {
-    fn is_empty(&self) -> bool {
-        self.read_lock().is_empty()
-    }
-
-    fn len(&self) -> Index {
-        self.read_lock().len()
-    }
-
-    fn get(&self, index: Index) -> T {
-        self.read_lock().get(index)
-    }
-
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        self.read_lock().get_many(indices)
-    }
-
-    /// Return all stored elements in a vector, whose index matches the StorageVec's.
-    /// It's the caller's responsibility that there is enough memory to store all elements.
-    fn get_all(&self) -> Vec<T> {
-        self.read_lock().get_all()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        self.write_lock().set(index, value)
-    }
-
-    /// set multiple elements.
-    ///
-    /// panics if key_vals contains an index not in the collection
-    ///
-    /// It is the caller's responsibility to ensure that index values are
-    /// unique.  If not, the last value with the same index will win.
-    /// For unordered collections such as HashMap, the behavior is undefined.
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        self.write_lock().set_many(key_vals)
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        self.write_lock().pop()
-    }
-
-    fn push(&mut self, value: T) {
-        self.write_lock().push(value)
-    }
-}
-
-impl<T: Serialize + DeserializeOwned> RustyLevelDbVec<T> {
-    // Return the key used to store the length of the persisted vector
-    pub fn get_length_key(key_prefix: u8) -> [u8; 2] {
-        RustyLevelDbVecPrivate::<T>::get_length_key(key_prefix)
-    }
-
-    /// Return the length at the last write to disk
-    pub fn persisted_length(&self) -> Index {
-        self.read_lock().persisted_length()
-    }
-
-    /// Return the level-DB key used to store the element at an index
-    pub fn get_index_key(&self, index: Index) -> [u8; 9] {
-        self.read_lock().get_index_key(index)
-    }
-
-    pub fn new(db: Arc<DB>, key_prefix: u8, name: &str) -> Self {
-        Self {
-            inner: Arc::new(RwLock::new(RustyLevelDbVecPrivate::<T>::new(
-                db, key_prefix, name,
-            ))),
-        }
-    }
-
-    /// Collect all added elements that have not yet bit persisted
-    pub fn pull_queue(&mut self, write_batch: &mut WriteBatch) {
-        self.write_lock().pull_queue(write_batch)
-    }
-
-    pub(crate) fn read_lock(&self) -> RwLockReadGuard<'_, RustyLevelDbVecPrivate<T>> {
-        self.inner.read().unwrap()
-    }
-
-    pub(crate) fn write_lock(&mut self) -> RwLockWriteGuard<'_, RustyLevelDbVecPrivate<T>> {
-        self.inner.write().unwrap()
-    }
-}
-
-pub struct OrdinaryVec<T>(Vec<T>);
-
-impl<T: Clone> StorageVec<T> for OrdinaryVec<T> {
-    fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    fn len(&self) -> Index {
-        self.0.len() as Index
-    }
-
-    fn get(&self, index: Index) -> T {
-        self.0[index as usize].clone()
-    }
-
-    fn get_many(&self, indices: &[Index]) -> Vec<T> {
-        indices
-            .iter()
-            .map(|index| self.0[*index as usize].clone())
-            .collect()
-    }
-
-    fn get_all(&self) -> Vec<T> {
-        self.0.clone()
-    }
-
-    fn set(&mut self, index: Index, value: T) {
-        // note: on 32 bit systems, this could panic.
-        self.0[index as usize] = value;
-    }
-
-    /// set multiple elements.
-    ///
-    /// panics if key_vals contains an index not in the collection
-    ///
-    /// It is the caller's responsibility to ensure that index values are
-    /// unique.  If not, the last value with the same index will win.
-    /// For unordered collections such as HashMap, the behavior is undefined.
-    fn set_many(&mut self, key_vals: impl IntoIterator<Item = (Index, T)>) {
-        for (index, value) in key_vals.into_iter() {
-            // note: on 32 bit systems, this could panic.
-            self.0[index as usize] = value;
-        }
-    }
-
-    fn pop(&mut self) -> Option<T> {
-        self.0.pop()
-    }
-
-    fn push(&mut self, value: T) {
-        self.0.push(value);
-    }
-}
+// We have split storage_vec into individual files, but for compatibility
+// we still keep everything in mod storage_vec.
+//
+// To accomplish that, we keep the sub modules private, and
+// add `pub use sub_module::*`.
+
+mod ordinary_vec;
+mod rusty_leveldb_vec;
+mod rusty_leveldb_vec_private;
+mod storage_vec_trait;
+
+pub use {ordinary_vec::*, rusty_leveldb_vec::*, storage_vec_trait::*};
 
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
 
+    use crate::storage::level_db::DB;
+
     use super::*;
+    use itertools::Itertools;
     use rand::{Rng, RngCore};
+    use std::collections::HashMap;
 
     use leveldb::{
         batch::{Batch, WriteBatch},
@@ -623,7 +127,7 @@ mod tests {
             RustyLevelDbVec::new(db, 0, "unit test vec 0");
         simple_prop(delegated_db_vec);
 
-        let ordinary_vec = OrdinaryVec::<[u8; 13]>(vec![]);
+        let ordinary_vec = OrdinaryVec::<[u8; 13]>::from(vec![]);
         simple_prop(ordinary_vec);
     }
